requirements/serve.txt
jinja2==3.0.0
Flask==2.0.1
koalas==1.8.1
raydp==0.3.0
pandas==1.1.4
pyspark==3.0.3
tox==3.14.6
python-dotenv==0.12.0
markdown==2.6.11
requests==2.25.1
alabaster==0.7.12
Babel==2.9.0
bleach==3.2.1
colorama==0.4.4
confluent-kafka==1.5.0
docutils==0.16
imagesize==1.2.0
MarkupSafe>=2.0.0
pkginfo==1.6.1
pycryptodome==3.9.9
Pygments==2.7.2
python-dateutil==2.8.1
pytz==2020.4
readme-renderer==28.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
twine==3.2.0
typing-extensions==3.7.4.3
webencodings==0.5.1
artifacts-keyring==0.3.1
sasl==0.2.1
thrift==0.13.0
thrift-sasl==0.3.0
PyHive==0.6.1
hdfs==2.5.8
SQLAlchemy==1.4.15
